Move balance power to a mod
